+++ src/main/java/com/twilio/rest/Domains.java
@@ -10,7 +10,6 @@ package com.twilio.rest;
 public enum Domains {
     ACCOUNTS("accounts"),
     API("api"),
-    AUTHY("authy"),
     AUTOPILOT("autopilot"),
     CHAT("chat"),
     CONVERSATIONS("conversations"),