+++ src/main/java/com/twilio/rest/supersim/v1/Sim.java
@@ -71,6 +71,7 @@ public class Sim extends Resource {
     }
 
     public enum StatusUpdate {
+        READY("ready"),
         ACTIVE("active"),
         INACTIVE("inactive");