+++ src/main/java/com/twilio/rest/verify/v2/service/WebhookFetcher.java
@@ -5,9 +5,9 @@
  *       /       /
  */
 
-package com.twilio.rest.authy.v1;
+package com.twilio.rest.verify.v2.service;
 
-import com.twilio.base.Deleter;
+import com.twilio.base.Fetcher;
 import com.twilio.exception.ApiConnectionException;
 import com.twilio.exception.ApiException;
 import com.twilio.exception.RestException;
@@ -22,36 +22,41 @@ import com.twilio.rest.Domains;
  * change. Use them with caution. If you currently do not have developer preview
  * access, please contact help@twilio.com.
  */
-public class ServiceDeleter extends Deleter<Service> {
+public class WebhookFetcher extends Fetcher<Webhook> {
+    private final String pathServiceSid;
     private final String pathSid;
 
     /**
-     * Construct a new ServiceDeleter.
+     * Construct a new WebhookFetcher.
      *
-     * @param pathSid A string that uniquely identifies this Service.
+     * @param pathServiceSid Service Sid.
+     * @param pathSid The unique string that identifies the resource to fetch
      */
-    public ServiceDeleter(final String pathSid) {
+    public WebhookFetcher(final String pathServiceSid,
+                          final String pathSid) {
+        this.pathServiceSid = pathServiceSid;
         this.pathSid = pathSid;
     }
 
     /**
-     * Make the request to the Twilio API to perform the delete.
+     * Make the request to the Twilio API to perform the fetch.
      *
      * @param client TwilioRestClient with which to make the request
+     * @return Fetched Webhook
      */
     @Override
     @SuppressWarnings("checkstyle:linelength")
-    public boolean delete(final TwilioRestClient client) {
+    public Webhook fetch(final TwilioRestClient client) {
         Request request = new Request(
-            HttpMethod.DELETE,
-            Domains.AUTHY.toString(),
-            "/v1/Services/" + this.pathSid + ""
+            HttpMethod.GET,
+            Domains.VERIFY.toString(),
+            "/v2/Services/" + this.pathServiceSid + "/Webhooks/" + this.pathSid + ""
         );
 
         Response response = client.request(request);
 
         if (response == null) {
-            throw new ApiConnectionException("Service delete failed: Unable to connect to server");
+            throw new ApiConnectionException("Webhook fetch failed: Unable to connect to server");
         } else if (!TwilioRestClient.SUCCESS.apply(response.getStatusCode())) {
             RestException restException = RestException.fromJson(response.getStream(), client.getObjectMapper());
             if (restException == null) {
@@ -60,6 +65,6 @@ public class ServiceDeleter extends Deleter<Service> {
             throw new ApiException(restException);
         }
 
-        return response.getStatusCode() == 204;
+        return Webhook.fromJson(response.getStream(), client.getObjectMapper());
     }
 }